Create serverstats
This could be used to simply audit changes to a server and keep track of files. And it could be used to delete files, exfiltrate data, monitor critical changes... 

Because the script only takes a few seconds to run ... it's an in and out affair. Again, this is for educational purposes.